Fix toolbar tests
